fix #31 css->scss, and explicitly check RAILS_RELATIVE_URL_ROOT
Issue was fontawesome and ionicons had hard coded paths to files
which were not getting replaced with paths to compiled assets, when
deploying in production mode.

Convert fontawesome and ionicons to scss, then use font-url instead
of url to locate the appropriate files with names name-some_sha256.
Additionally use map in config.ru to respect baseurl when serving
with puma.

Tested in development, and production mode, only difference with
production is satisfying the requirement that secrets be set.