Fix #53 by adding a LaTeX compiled formula of compliance
Add as a png because currently the treemap wipes all svgs on the screen then
redraws, this includes svg's outside of scope. So we cannot insert an SVG
formula yet. Need to ask Rob if it is okay to change the treemap
selectAll('svg') line in treemap or if theres another reason for it.

Signed-off-by: Luke Malinowski <lmalinowski@mitre.org>